requirements.txt
Flask>=2.2.0
Werkzeug>=2.2.0
python-dotenv>=0.20.0